requirements.txt
django-graphql-jwt
graphene-django